# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Torvix is an open-source FinOps platform for cloud cost visibility, anomaly detection, forecasting, and unused-resource detection across cloud providers.",
        "title": "Torvix API",
        "contact": {
            "name": "Torvix Team"
        },
        "version": "0.7.0"
    },
    "host": "localhost:8080",
    "basePath": "",
    "paths": {
        "/api/v1/analytics/anomalies": {
            "get": {
                "description": "Analyses cost data over a time range and returns any detected anomalies with severity, z-score, and deviation metrics.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Detect cost anomalies",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2025-01-01",
                        "description": "Start date (YYYY-MM-DD). Defaults to 30 days ago.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-01-31",
                        "description": "End date (YYYY-MM-DD). Defaults to today.",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detected anomalies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_crypticani_torvix_internal_domain.Anomaly"
                            }
                        }
                    },
                    "500": {
                        "description": "Detection failed",
                        "schema": {
                            "$ref": "#/definitions/internal_ports_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/forecast": {
            "get": {
                "description": "Generates a 7-day cost forecast based on historical data within the specified time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forecasting"
                ],
                "summary": "Get cost forecast",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2025-01-01",
                        "description": "Start date (YYYY-MM-DD). Defaults to 30 days ago.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-01-31",
                        "description": "End date (YYYY-MM-DD). Defaults to today.",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Forecast data points",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_crypticani_torvix_internal_domain.ForecastPoint"
                            }
                        }
                    },
                    "500": {
                        "description": "Forecast generation failed",
                        "schema": {
                            "$ref": "#/definitions/internal_ports_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/summary": {
            "get": {
                "description": "Returns aggregated cost data over a time range, grouped by the specified window (daily, weekly, or monthly).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get cost summary",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2025-01-01",
                        "description": "Start date (YYYY-MM-DD). Defaults to 30 days ago.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-01-31",
                        "description": "End date (YYYY-MM-DD). Defaults to today.",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "daily",
                            "weekly",
                            "monthly"
                        ],
                        "type": "string",
                        "default": "daily",
                        "description": "Aggregation window: daily, weekly, or monthly.",
                        "name": "window",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cost aggregation results",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_crypticani_torvix_internal_domain.AggregatedCost"
                            }
                        }
                    },
                    "500": {
                        "description": "Aggregation failed",
                        "schema": {
                            "$ref": "#/definitions/internal_ports_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ingest": {
            "post": {
                "description": "Queues collection of billing data from all enabled cloud providers within the configured rolling ingestion window and returns immediately. Completion is available through the returned status URL and enabled alerting targets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ingestion"
                ],
                "summary": "Trigger billing data ingestion",
                "responses": {
                    "202": {
                        "description": "Ingestion queued for background processing",
                        "schema": {
                            "$ref": "#/definitions/internal_ports_http.IngestAcceptedResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Ingestion service unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_ports_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ingest/status/{job_id}": {
            "get": {
                "description": "Returns the current or completed status for a background ingestion job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ingestion"
                ],
                "summary": "Get ingestion job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingestion job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingestion job status",
                        "schema": {
                            "$ref": "#/definitions/internal_ports_http.IngestionJobResponse"
                        }
                    },
                    "404": {
                        "description": "Ingestion job not found",
                        "schema": {
                            "$ref": "#/definitions/internal_ports_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_crypticani_torvix_internal_domain.AggregatedCost": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_crypticani_torvix_internal_domain.Provider"
                },
                "service": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "number"
                },
                "window_end": {
                    "type": "string"
                },
                "window_start": {
                    "type": "string"
                }
            }
        },
        "github_com_crypticani_torvix_internal_domain.Anomaly": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "actual": {
                    "type": "number"
                },
                "baseline": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "compartment_id": {
                    "type": "string"
                },
                "compartment_name": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "moving_average_delta": {
                    "type": "number"
                },
                "percent_deviation": {
                    "type": "number"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_crypticani_torvix_internal_domain.Provider"
                },
                "region": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "z_score": {
                    "type": "number"
                }
            }
        },
        "github_com_crypticani_torvix_internal_domain.ForecastPoint": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "confidence_high": {
                    "type": "number"
                },
                "confidence_low": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "forecast_cost": {
                    "type": "number"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_crypticani_torvix_internal_domain.Provider"
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "github_com_crypticani_torvix_internal_domain.Provider": {
            "type": "string",
            "enum": [
                "aws",
                "azure",
                "gcp",
                "oci"
            ],
            "x-enum-varnames": [
                "ProviderAWS",
                "ProviderAzure",
                "ProviderGCP",
                "ProviderOCI"
            ]
        },
...
        "internal_ports_http.ErrorResponse": {
            "description": "Error response returned when an operation fails.",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
    }
}